+++ package/example/lib/pages/fullscreen_example_page.dart
@@ -33,8 +33,8 @@ class _FullscreenExamplePageState extends State<FullscreenExamplePage> {
   // );
 
   final MeeduPlayerController _meeduPlayerController = MeeduPlayerController(
-    colorTheme: Colors.blue,
-  );
+      colorTheme: Colors.blue,
+      enabledButtons: EnabledButtons(rewindAndfastForward: false));
   ValueNotifier<int> currentIndex = ValueNotifier(0);
   DataSource? _dataSource;