+++ package/example/lib/pages/listview_example.dart
@@ -44,6 +44,8 @@ class _VideoItemState extends State<VideoItem>
     screenManager: const ScreenManager(orientations: [
       DeviceOrientation.portraitUp,
     ]),
+    enabledControls: EnabledControls(doubleTapToSeek: false)
+
   );
 
   final ValueNotifier<bool> _visible = ValueNotifier(true);