+++ package/example/lib/pages/m3u8_page_example.dart
@@ -214,8 +214,9 @@ class M3u8ExamplePage extends StatefulWidget {
 
 class _M3u8ExamplePageState extends State<M3u8ExamplePage> {
   final _controller = MeeduPlayerController(
-    screenManager: const ScreenManager(forceLandScapeInFullscreen: false),
-  );
+      screenManager: const ScreenManager(forceLandScapeInFullscreen: false),
+      enabledButtons: EnabledButtons(rewindAndfastForward: false),
+      responsive: Responsive(buttonsSizeRelativeToScreen: 3));
   String fileName = "";
   List<Quality> _qualities = [];